V3 - GPCalendar (Shell completed, API completed, README updated)

+++ include/GPCalendar/Model/rendezvous.h
@@ -23,7 +23,7 @@ typedef struct{
 RendezVous * CreerRendezVous(int an, int mois, int jour, double heure_debut, double duree, char * lieu, Patient * patient, Medecin * medecin, char * motif);
 void FreeRendezVous(RendezVous * rdv);
 int EqualsRendezVous(RendezVous * rdv1, RendezVous * rdv2);
-RendezVous * DeplacerRendezVous(RendezVous * rdv, int n_an, int n_mois, int n_jour, double n_heure_debut, int n_duree);
+RendezVous * DeplacerRendezVous(RendezVous * rdv, int n_an, int n_mois, int n_jour, double n_heure_debut, double n_duree);
 
 void getInfosRendezVous(char* infos, RendezVous* rdv);
